requirements.txt
geoalchemy2
Pillow
sqlalchemy
